+++ src/harlequin_duckdb/__init__.py
@@ -1,0 +1,3 @@
+from harlequin_duckdb.adapter import DuckDbAdapter as DuckDbAdapter
+from harlequin_duckdb.cli_options import DUCKDB_OPTIONS as DUCKDB_OPTIONS
+from harlequin_duckdb.copy_formats import DUCKDB_COPY_FORMATS as DUCKDB_COPY_FORMATS