+++ tests/data/functional_tests/small/schema_sqlite.sql
@@ -1,0 +1,5 @@
+CREATE TABLE drivers(driverId BIGINT, driverRef VARCHAR, number VARCHAR, code VARCHAR, forename VARCHAR, surname VARCHAR, dob VARCHAR, nationality VARCHAR, url VARCHAR);
+
+
+
+