+++ tests/data/functional_tests/tiny/schema_sqlite.sql
@@ -1,0 +1,8 @@
+
+
+
+CREATE TABLE foo(foo_col INTEGER);
+
+
+
+